refactor(18150): rename and relocate hte store hook
