+++ hivemq-edge/src/frontend/src/modules/EdgeVisualisation/hooks/useWorkspaceStore.ts
